team1Python/requirements.txt
biopython==1.79
certifi
lxml
ipython
pandas
requests
urllib3